move manual filtering examples to a separate page
